Prevent removal of command containing a mention

+++ NyaChan.py
@@ -48,11 +48,25 @@ async def on_command_error(ctx, error):
     else:
         await ctx.author.send(
             "{}, error```py\n{}: {}\n```".format(ctx.message.author.mention, type(error).__name__, str(error)))
+    has_mention = False
+    for arg in ctx.args:
+        if '@' in str(arg):
+            has_mention = True
+            break
+    if has_mention is True:
+        return False
     await ctx.message.delete()
 
 
 @bot.event
 async def on_command_completion(ctx):
+    has_mention = False
+    for arg in ctx.args:
+        if '@' in str(arg):
+            has_mention = True
+            break
+    if has_mention is True:
+        return False
     await ctx.message.delete()
